requirements.txt
# Core
pydantic>=2.0
pyyaml>=6.0
numpy>=1.24
scipy>=1.10
Pillow>=10.0

# CLIP and image processing
open_clip_torch>=2.24
clip-retrieval>=2.6
imagehash>=4.3

# HuggingFace
transformers>=4.40
diffusers>=0.27
datasets>=2.18
accelerate>=0.28
safetensors>=0.4

# LLM inference
vllm>=0.4.0
openai>=1.12

# ML utilities
torch>=2.0
scikit-learn>=1.3

# Data processing
tqdm>=4.65
